Rename of test to reflect that it is testing the Fake user login
